Struct libtrig::Angle2D

#[repr(C)]
pub struct Angle2D(/* private fields */);

A wrapper around a float value that represents an angle.

It can be created from either radians or degrees, and can be converted to either radians or degrees.

Implementations

impl Angle2D

pub const fn deg_to_rad(deg: Degree64) -> Radian64

Converts the given deg value to radians.

pub const fn rad_to_deg(rad: Radian64) -> Degree64

Converts the given rad value to degrees.

impl Angle2D

pub const fn new(value: RadianOrDegree64, is_radians: bool) -> Self

Creates a new Angle2D from the given value and is_radians flag.

pub const fn from_radians(value: Radian64) -> Self

Creates a new Angle2D from the given value in radians.

pub const fn from_degrees(value: Degree64) -> Self

Creates a new Angle2D from the given value in degrees.

pub const fn zero() -> Self

Creates a new Angle2D with a value of 0.0 radians.

pub const fn zero_deg() -> Self

Creates a new Angle2D with a value of 0.0 degrees.

pub const fn to_radians(&self) -> Radian64

Returns the value of the angle in radians.

pub const fn to_degrees(&self) -> Degree64

Returns the value of the angle in degrees.

pub const fn set_self_radians(&mut self)

If the angle is in degrees, converts it to radians.

pub const fn set_self_degrees(&mut self)

If the angle is in radians, converts it to degrees.

pub const unsafe fn modify_self_mode_unchecked(&mut self, mode: bool)

Sets the mode of self.

If mode is true, then the angle is in radians. If mode is false, then the angle is in degrees.

This is unsafe because it does not convert the internal value.

pub const unsafe fn modify_self_value_unchecked( &mut self, value: RadianOrDegree64 )

Sets the value of self.

This is unsafe because it does not convert the internal value.