Struct libtrig::Coord2D

#[repr(C)]
pub struct Coord2D { /* private fields */ }

A wrapper around a float value that represents a 2D coordinate.

Implementations

impl Coord2D

pub const fn new(x: Float64, y: Float64) -> Self

Creates a new Vec2D from the given x and y values.

pub const fn origin() -> Self

Creates a new Vec2D located at the origin.

pub fn rotate_by(&self, angle: Angle2D) -> Self

Rotates the vector by the given angle in radians.

pub fn angle(&self) -> Angle2D

Returns the angle of the vector in radians.

pub fn inverse(&self) -> Self

Returns the inverted position.

Same as rotating the position by 180 degrees.

pub const fn x(&self) -> Float64

Returns the x coordinate.

pub const fn y(&self) -> Float64

Returns the y coordinate.