Struct libtrig::u3

#[repr(C)]
pub struct u3(/* private fields */);

The 3-bit unsigned integer type.

Implementations

impl u3

pub const MAX: Self = _

Maximum value of u3. (7)

pub const MIN: Self = _

Minimum value of u3. (0)

pub const ZERO: Self = Self::MIN

Zero value of u3. (0)

pub const ONE: Self = _

One value of u3. (1)

pub const TWO: Self = _

Two value of u3. (2)

pub const THREE: Self = _

Three value of u3. (3)

pub const FOUR: Self = _

Four value of u3. (4)

pub const FIVE: Self = _

Five value of u3. (5)

pub const SIX: Self = _

Six value of u3. (6)

pub const SEVEN: Self = Self::MAX

Seven value of u3. (7)

pub const fn new(ones: bool, twos: bool, fours: bool) -> Self

Create a new u3 from three bits.

pub const fn tou8(self) -> u8

Convert a u3 to a u8.

pub const unsafe fn fromu8(u: u8) -> Self

Convert a u8 to a u3.

Safety

This uses core::hint::unreachable_unchecked to enable optimizations, so it is unsafe to use in debug mode. Instead, use u3::fromu8.

pub const fn tou2(self) -> u2

Convert a u3 to a u2.

pub const fn fromu2(u: u2) -> Self

Convert a u2 to a u3.

pub const fn zero(self) -> bool

Returns true if self is 0.

pub const fn one(self) -> bool

Returns true if self is 1.

pub const fn two(self) -> bool

Returns true if self is 2.

pub const fn three(self) -> bool

Returns true if self is 3.

pub const fn four(self) -> bool

Returns true if self is 4.

pub const fn five(self) -> bool

Returns true if self is 5.

pub const fn six(self) -> bool

Returns true if self is 6.

pub const fn seven(self) -> bool

Returns true if self is 7.

pub const fn bitnot(self) -> Self

Inverts all bits in a u3.

pub const unsafe fn unchecked_add(self, rhs: Self) -> Self

Adds two u3s.

Safety

This function is unsafe because it may overflow, therefore causing undefined behavior. The developer must ensure that the result is in the range 0..=7.

pub const unsafe fn unchecked_sub(self, rhs: Self) -> Self

Subtracts two u3s.

Safety

This function is unsafe because it may overflow, therefore causing undefined behavior. The developer must ensure that the result is in the range 0..=7.

pub const unsafe fn unchecked_mul(self, rhs: Self) -> Self

Performs an unchecked multiplication of two u3s.

Safety

This function is unsafe because it may overflow, therefore causing undefined behavior. The developer must ensure that the result is in the range 0..=7.

pub const unsafe fn unchecked_div(self, rhs: Self) -> Self

Performs an unchecked division of two u3s.

Safety

This function is unsafe because it may overflow, therefore causing undefined behavior. The developer must ensure that the result is in the range 0..=7.